// src/generated/opentofu-api.ts
export interface operations {
// ...
  GetProviderDocItem: {
    parameters: {
      path: {
        /** Namespace of the provider, all lower case. */
        namespace: string;
        /** Name of the provider, all lower case. */
        name: string;
        /** Version number of the provider with the "v" prefix. */
        version: string;
        /** The kind of document to fetch. */
        kind: "resource" | "datasource" | "function" | "guide";
        /** The name of the document without the .md suffix. */
        document: string;
      };
    };
    responses: {
      /** The contents of the document. */
      200: {
        schema: unknown;
      };
    };
  };
// ...
}